added new config to basic starter
