--- examples/basic/.tina/components/provider/index.jsx
@@ -1,15 +1,0 @@
-import dynamic from 'next/dynamic'
-const TinaProvider = dynamic(() => import('./_TinaProvider'), { ssr: false })
-import { TinaEditProvider } from 'tinacms/dist/edit-state'
-
-const DynamicTina = ({ children }) => {
-  return (
-    <>
-      <TinaEditProvider editMode={<TinaProvider>{children}</TinaProvider>}>
-        {children}
-      </TinaEditProvider>
-    </>
-  )
-}
-
-export default DynamicTina
