+++ examples/basic/package.json
@@ -17,6 +17,7 @@
   },
   "devDependencies": {
     "@tinacms/cli": "workspace:*",
+    "@types/node": "^17.0.21",
     "eslint": "7",
     "eslint-config-next": "12.0.3"
   }
